requirements.txt
altgraph==0.17.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
future==0.18.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
pefile==2022.5.30
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pywin32-ctypes==0.2.0
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
